Fix ability bar refresh while grappling.
Enabling LoS updates while the current unit is performing a grapple targeting action causes
some weirdness with the ability bar closing and reopening each time the target location moves.
This is due to a call to rebuild abilities on each visibility update, and visibility updates
occur when moving the helpers around. To fix this, avoid an ability refresh if the current
unit is perfoming a grapple action. This should hopefully not occur in any other scenario.